refactor: relocate analyzeworkflowsview to the shared package and drop the side dispatcher (#1606)
