+++ app/apis/catalog/brc-analytics-catalog/common/utils.ts
@@ -1,7 +1,3 @@
-import {
-  ORGANISM_PLOIDY,
-  WORKFLOW_PLOIDY,
-} from "@repo/shared/apis/schema-types";
 import { sanitizeEntityId } from "@repo/shared/apis/utils";
 import { formatTrsId } from "@repo/shared/workflow/utils";
 import { type WorkflowEntity } from "@site-config/brc-analytics/local/index/workflow/types";
@@ -39,26 +35,4 @@ export function getGenomeOrganismId(genome: BRCDataCatalogGenome): string {
  */
 export function getWorkflowId(workflow: WorkflowEntity): string {
   return formatTrsId(workflow.trsId);
-}
-
-/**
- * Get whether a given workflow ploidy is compatible with a given organism ploidy.
- * @param workflowPloidy - Workflow ploidy.
- * @param organismPloidy - Organism ploidy.
- * @returns boolean indicating whether the given ploidy values are compatible.
- */
-export function workflowPloidyMatchesOrganismPloidy(
-  workflowPloidy: WORKFLOW_PLOIDY,
-  organismPloidy: ORGANISM_PLOIDY
-): boolean {
-  switch (workflowPloidy) {
-    case WORKFLOW_PLOIDY.ANY:
-      return true;
-    case WORKFLOW_PLOIDY.DIPLOID:
-      return organismPloidy === ORGANISM_PLOIDY.DIPLOID;
-    case WORKFLOW_PLOIDY.HAPLOID:
-      return organismPloidy === ORGANISM_PLOIDY.HAPLOID;
-    case WORKFLOW_PLOIDY.POLYPLOID:
-      return organismPloidy === ORGANISM_PLOIDY.POLYPLOID;
-  }
 }
