--- app/views/AnalyzeWorkflowsView/types.ts
@@ -1,3 +1,0 @@
-export interface Props {
-  entityId: string;
-}
