--- app/views/EntityView/assembly/components/Side/types.ts
@@ -1,5 +1,0 @@
-import { type Assembly } from "@/views/WorkflowInputsView/types";
-
-export interface Props {
-  assembly: Assembly;
-}
