+++ app/views/WorkflowsView/components/Workflows/components/Card/card.tsx
@@ -1,4 +1,3 @@
-import { CHIP_PROPS } from "@/views/AnalyzeWorkflowsView/components/Main/components/Accordion/constants";
 import { RoundedPaper } from "@databiosphere/findable-ui/lib/components/common/Paper/components/RoundedPaper/roundedPaper";
 import {
   ANCHOR_TARGET,
@@ -18,6 +17,7 @@ import {
 } from "@mui/material";
 import { TruncatedText } from "@repo/shared/components/TruncatedText/truncatedText";
 import { ROUTES } from "@repo/shared/routes/constants";
+import { CHIP_PROPS } from "@repo/shared/views/AnalyzeWorkflowsView/components/Main/components/Accordion/constants";
 import Link from "next/link";
 import { type JSX } from "react";
 import { StyledCard } from "./card.styles";
