+++ packages/shared/views/AnalyzeWorkflowsView/components/Main/components/Accordion/components/Workflow/workflow.styles.ts
