+++ packages/shared/views/AnalyzeWorkflowsView/components/Top/types.ts
@@ -1,0 +1,6 @@
+import { type AssemblyContract } from "@repo/shared/apis/types";
+
+export interface Props {
+  assembly: AssemblyContract;
+  entityId: string;
+}
