+++ tests/views/analyzeWorkflowsView/buildAssemblyWorkflows.scope.test.ts
@@ -1,11 +1,11 @@
 import type { BRCDataCatalogGenome } from "@/apis/catalog/brc-analytics-catalog/common/entities";
-import { buildAssemblyWorkflows } from "@/views/AnalyzeWorkflowsView/components/Main/utils";
 import {
   ORGANISM_PLOIDY,
   WORKFLOW_PLOIDY,
   WORKFLOW_SCOPE,
 } from "@repo/shared/apis/schema-types";
 import type { WorkflowCategory } from "@repo/shared/apis/workflow";
+import { buildAssemblyWorkflows } from "@repo/shared/views/AnalyzeWorkflowsView/components/Main/utils";
 import { DIFFERENTIAL_EXPRESSION_ANALYSIS } from "@repo/shared/workflow/differentialExpressionAnalysis";
 import { WorkflowCategoryId } from "../../../catalog/schema/generated/schema";
